Update readme instructions for css and images
The instructions should be changed to properly reflect the way css and
images are handled in the configuration.

At the moment the instructions reference css and png files being
handled via templateFormats, but in reality they're configured to be
located in special css and img directories and handled via
addPassthroughCopy().